refactor: Code architecture
