+++ dristi/lib/src/core/widgets/error_screen.dart
@@ -1,4 +1,4 @@
-import 'package:dristi/src/core/assets/assets.dart';
+import 'package:dristi/src/core/constants/app_assets.dart';
 import 'package:dristi/src/core/constants/app_values.dart';
 import 'package:dristi/src/core/theme/colors.dart';
 import 'package:dristi/src/core/theme/text_styles.dart';
