+++ dristi/lib/src/features/home/categories/data/data_sources/categories_data_source_impl.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:dristi/src/core/base/base_remote_source.dart';
-import 'package:dristi/src/core/network/api_end_points.dart';
+import 'package:dristi/src/core/constants/app_end_points.dart';
 import 'package:dristi/src/core/network/dio_provider.dart';
 import 'package:dristi/src/features/home/categories/data/data_sources/categories_data_source.dart';
 import 'package:dristi/src/features/home/categories/data/model/categories_response_model.dart';
