+++ dristi/lib/src/features/home/top_destinations/data/data_sources/top_destination_data_source_impl.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:dio/dio.dart';
-import 'package:dristi/src/core/assets/assets.dart';
+import 'package:dristi/src/core/constants/app_assets.dart';
 import 'package:dristi/src/core/loggers/logger.dart';
 import 'package:dristi/src/features/home/top_destinations/data/data_sources/top_destination_data_source.dart';
 import 'package:dristi/src/features/home/top_destinations/data/model/top_destination_response_model.dart';
